Run  worker code in a child process
