+++ package.json
@@ -26,6 +26,7 @@
     "watch": "nodemon --watch ./src --exec npm run serve",
     "watch:all": "concurrently --prefix '{pid}-{name}' --names 'API,WORKERS' --kill-others 'npm run watch' 'npm run worker:watch'",
     "worker:serve": "NODE_ENV=development babel-node -r dotenv/config ./src/worker.js --source-maps inline",
+    "worker-test:serve": "NODE_ENV=test babel-node -r dotenv/config ./src/worker.js --source-maps inline",
     "worker:start": "cross-env NODE_ENV=production node ./build/worker.js",
     "worker:watch": "nodemon --watch ./src --exec npm run worker:serve"
   },
@@ -40,7 +41,7 @@
     "babel-eslint": "^10.1.0",
     "babel-plugin-add-module-exports": "^1.0.4",
     "concurrently": "^5.3.0",
-    "cross-env": "^7.0.2",
+    "cross-env": "^7.0.3",
     "eslint": "^7.12.1",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-prettier": "^3.1.4",
